+++ heinlein/manager/dataManger.py
@@ -4,6 +4,7 @@ from abc import ABC, abstractmethod
 from importlib import import_module
 from inspect import getmembers, isclass, isfunction
 from pathlib import Path
+from types import ModuleType
 
 from cacheout import LRUCache
 from godata import create_project, has_collection, has_project, load_project
@@ -15,6 +16,29 @@ from heinlein.region.base import BaseRegion
 from heinlein.utilities import warning_prompt_tf
 
 logger = logging.getLogger("manager")
+
+
+known_datasets = ["des", "cfht", "hsc", "ms"]
+
+
+def get_external_implementation(name: str) -> ModuleType:
+    """
+    Checks to see if a custom implementation exists for a given dataset. These have
+    to be installed separately from the main package. Prompts the user to install
+    the package if it is not found, but is known.
+    """
+    if name in known_datasets:
+        # try to import it
+        module_name = "heinlein_" + name
+        try:
+            module = import_module(module_name)
+        except ImportError:
+            print(
+                f"Dataset `{name}` is a known dataset, but needs to be installed"
+                f" separately. You can install it with `heinlein install {name}`."
+            )
+            exit()
+        return getattr(module, name)
 
 
 def check_overload(f):
@@ -96,7 +120,7 @@ class DataManager(ABC):
             try:
                 # Find the external implementation for this dataset, if it exists.
                 cfg = self.config.get("config")
-                self.external = import_module(f".{cfg['slug']}", "heinlein.dataset")
+                self.external = get_external_implementation(cfg["slug"])
             except KeyError:
                 self.external = None
             self._initialize_external_implementation()
